requirements.txt
anyio==4.7.0
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.7
Django==5.1.4
django-cors-headers==4.6.0
graphql-core==3.2.5
h11==0.14.0
httptools==0.6.4
idna==3.10
msgpack==1.1.0
mysqlclient==2.2.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
strawberry-graphql==0.254.0
strawberry-graphql-django==0.51.0
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
